+++ app/locales/zh/translations.ts
@@ -629,4 +629,8 @@ export default {
     paginatedRelation: {
         error: 'There was an error loading this list.',
     },
+    cookieBanner: {
+        cookieWarning: 'This website relies on cookies to help provide a better user experience. By clicking Accept or continuing to use the site, you agree. For more information, see our <a href="https://github.com/CenterForOpenScience/cos.io/blob/master/PRIVACY_POLICY.md">Privacy Policy</a> and information on <a href="https://github.com/CenterForOpenScience/cos.io/blob/master/PRIVACY_POLICY.md#7-types-of-information-we-collect">cookie use</a>.',
+        accept: 'Accept',
+    },
 };