+++ CHANGELOG.md
@@ -10,11 +10,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - `region`
 - Feature Flags:
     - `storage_i18n` - enable region selector on project creation
+- `keen` metrics adapter
 
 ### Changed
 - Models:
     - `node` - add `region` relationship
     - `user` - add `defaultRegion` relationship
+- Services:
+    - `analytics` - support multiple metrics adapters
 
 ## [0.6.0] - 2018-07-13
 ### Added