+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - `region`
 - Feature Flags:
     - `storage_i18n` - enable region selector on project creation
+- Components:
+    - `cookie-banner` - Notify logged-out users the site uses cookies
   - Assets:
     - Images for home page
     - Images for dashboard
@@ -21,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - `user` - add `defaultRegion` relationship
 - Services:
     - `analytics` - support multiple metrics adapters
+- Add `join-osf-banner` to `guid-node` pages.
 
 ## [0.6.0] - 2018-07-13
 ### Added