+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Assets:
     - Images for home page
     - Images for dashboard
+- Tests
+    - More dashboard application tests
 - `keen` metrics adapter
 - CSRF support on all ajax requests