+++ CHANGELOG.md
@@ -28,6 +28,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - `analytics` - support multiple metrics adapters
 - Add `join-osf-banner` to `guid-node` pages.
 
+## [0.6.1] - 2018-07-31
+### Changed
+- local storage key used for `ember-simple-auth`
+
 ## [0.6.0] - 2018-07-13
 ### Added
 - Models: